Refactor Logging.
