+++ backend/src/Logitar.Portal.Infrastructure/DependencyInjectionExtensions.cs
@@ -41,9 +41,9 @@ namespace Logitar.Portal.Infrastructure
         .AddQueriers()
         .AddSingleton<IPasswordService, PasswordService>()
         .AddSingleton<ITemplateCompiler, TemplateCompiler>()
-        .AddScoped<IInternalLoggingContext, InternalLoggingContext>()
         .AddScoped<IJwtBlacklist, JwtBlacklist>()
         .AddScoped<ILoggingContext, LoggingContext>()
+        .AddScoped<ILoggingService, LoggingService>()
         .AddScoped<IRepository, Repository>()
         .AddScoped<IRequestPipeline, RequestPipeline>()
         .AddScoped<ISecurityTokenService, JwtService>()
